api/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --no-emit-index-url
#
annotated-types==0.5.0
    # via pydantic
anyio==3.7.1
    # via
    #   fastapi
    #   starlette
    #   watchfiles
click==8.1.7
    # via uvicorn
covjson-pydantic @ git+https://github.com/KNMI/covjson-pydantic@KDP-1761
    # via -r requirements.in
fastapi==0.103.1
    # via -r requirements.in
grpcio==1.58.0
    # via grpcio-tools
grpcio-tools==1.58.0
    # via -r requirements.in
gunicorn==21.2.0
    # via -r requirements.in
h11==0.14.0
    # via uvicorn
httptools==0.6.0
    # via uvicorn
idna==3.4
    # via anyio
packaging==23.1
    # via gunicorn
protobuf==4.24.3
    # via grpcio-tools
pydantic==2.3.0
    # via
    #   covjson-pydantic
    #   fastapi
pydantic-core==2.6.3
    # via pydantic
python-dotenv==1.0.0
    # via uvicorn
pyyaml==6.0.1
    # via uvicorn
sniffio==1.3.0
    # via anyio
starlette==0.27.0
    # via fastapi
typing-extensions==4.7.1
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
uvicorn[standard]==0.23.2
    # via -r requirements.in
uvloop==0.17.0
    # via uvicorn
watchfiles==0.20.0
    # via uvicorn
websockets==11.0.3
    # via uvicorn

# The following packages are considered to be unsafe in a requirements file:
# setuptools
